Refactored business layer
