+++ ExamenPGG.Business.Tests/PlayerTests.cs
@@ -1,5 +1,5 @@
-﻿using ExamenPGG.Business._00_Interfaces;
-using ExamenPGG.Business._01_Classes;
+﻿using ExamenPGG.Business._01_Classes;
+using ExamenPGG.Business.Player;
 
 namespace ExamenPGG.Business.Tests
 {
