+++ ExamenPGG.Business/Dice.cs
@@ -1,6 +1,4 @@
-﻿using ExamenPGG.Business._00_Interfaces;
-
-namespace ExamenPGG.Business._01_Classes
+﻿namespace ExamenPGG.Business
 {
     public class Dice : IDice
     {
